Update README.md
It took me some time to realize that i did not switch from OS-mode to Run-mode. What puzzled me most was that the MQTT connection was "online", but no messages were send. (I also switched from PicoW to ESP... before realizing the mistake was on the configuration side.)

Therefore I tried to clarified the README 😄